Move `MeasureUnitParser` to `measure` crate
