+++ components/experimental/src/units/converter_factory.rs
@@ -2,7 +2,8 @@
 // called LICENSE at the top level of the ICU4X source tree
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
-use crate::measure::measureunit::{MeasureUnit, MeasureUnitParser};
+use crate::measure::measureunit::MeasureUnit;
+use crate::measure::parser::MeasureUnitParser;
 use crate::units::provider::{self, MeasureUnitItem};
 use crate::units::ratio::IcuRatio;
 use crate::units::{
